+++ elastalert/loaders.py
@@ -8,6 +8,7 @@ import sys
 import jsonschema
 import yaml
 import yaml.scanner
+from elastalert.alerters.flashduty import FlashdutyAlerter
 from jinja2 import Environment
 from jinja2 import FileSystemLoader
 from jinja2 import Template
@@ -146,6 +147,7 @@ class RulesLoader(object):
         'indexer': IndexerAlerter,
         'matrixhookshot': MatrixHookshotAlerter,
         'yzj': YzjAlerter,
+        'flashduty': FlashdutyAlerter,
         'smseagle': SMSEagleAlerter
     }